+++ src/funding/venmo/config.test.js
@@ -2,6 +2,8 @@
 
 import { describe, expect } from "vitest";
 
+import { BUTTON_FLOW } from "../../constants";
+
 import { getVenmoConfig } from "./config";
 
 describe("Venmo eligibility", () => {
@@ -10,6 +12,7 @@ describe("Venmo eligibility", () => {
     components: ["buttons"],
     fundingEligibility: {},
     wallet: expect.any,
+    flow: BUTTON_FLOW.PURCHASE,
   };
   const venmoConfig = getVenmoConfig();
 
@@ -43,6 +46,27 @@ describe("Venmo eligibility", () => {
 
   test("should be eligible if neither a shipping callback nor displayOnly is present", () => {
     const isVenmoEligible = venmoConfig.eligible?.(baseEligibilityProps);
+
+    expect(isVenmoEligible).toEqual(true);
+  });
+
+  test("should not be eligible if flow is VAULT_WITHOUT_PURCHASE and venmoVaultWithoutPurchase is false", () => {
+    const isVenmoEligible = venmoConfig.eligible?.({
+      ...baseEligibilityProps,
+      flow: BUTTON_FLOW.VAULT_WITHOUT_PURCHASE,
+    });
+
+    expect(isVenmoEligible).toEqual(false);
+  });
+
+  test("should be eligible if flow is VAULT_WITHOUT_PURCHASE and venmoVaultWithoutPurchase is true", () => {
+    const isVenmoEligible = venmoConfig.eligible?.({
+      ...baseEligibilityProps,
+      flow: BUTTON_FLOW.VAULT_WITHOUT_PURCHASE,
+      experiment: {
+        venmoVaultWithoutPurchase: true,
+      },
+    });
 
     expect(isVenmoEligible).toEqual(true);
   });